Extract a shared ShimmerBox composable
Consolidate the three identical alpha-pulse ShimmerBox copies (the new one in
PageRsRow plus the pre-existing ones in PostRsListItem and PostRsSettingsScreen)
into a single internal composable in postsrs/screens.
