+++ WordPress/src/main/java/org/wordpress/android/ui/postsrs/screens/ShimmerBox.kt
@@ -1,0 +1,37 @@
+package org.wordpress.android.ui.postsrs.screens
+
+import androidx.compose.animation.core.RepeatMode
+import androidx.compose.animation.core.animateFloat
+import androidx.compose.animation.core.infiniteRepeatable
+import androidx.compose.animation.core.rememberInfiniteTransition
+import androidx.compose.animation.core.tween
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.Modifier
+
+/**
+ * A simple loading placeholder that pulses its alpha. Shared by the wp-rs posts and
+ * pages list/settings screens. (The stats screens use a separate gradient-sweep
+ * shimmer in [org.wordpress.android.ui.newstats.util.ShimmerBox].)
+ */
+@Composable
+internal fun ShimmerBox(modifier: Modifier = Modifier) {
+    val infiniteTransition = rememberInfiniteTransition(label = "shimmer")
+    val alpha by infiniteTransition.animateFloat(
+        initialValue = 0.06f,
+        targetValue = 0.14f,
+        animationSpec = infiniteRepeatable(
+            animation = tween(durationMillis = 800),
+            repeatMode = RepeatMode.Reverse
+        ),
+        label = "shimmerAlpha"
+    )
+    Box(
+        modifier = modifier.background(
+            MaterialTheme.colorScheme.onSurface.copy(alpha = alpha)
+        )
+    )
+}
